apps/tts-server/pyproject.toml
[project]
name = "tts-server"
version = "0.2.0"
description = "统一语音服务 - 支持 ASR + TTS + WebSocket，本地推理和云端 API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # TTS 核心
    "genie-tts>=2.0.2",

    # Web 框架
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",

    # 配置管理
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",

    # HTTP 客户端（云端 API）
    "httpx>=0.27.2",
    "aiohttp>=3.10.0",

    # 音频处理
    "numpy>=2.4.1",
    "soundfile>=0.12.0",
]

[project.optional-dependencies]
# 本地 ASR - FunASR（阿里巴巴）
# ⚠️ 仅限 Linux/macOS - Windows 上存在以下兼容性问题：
#   - llvmlite 旧版本不支持 Python 3.12+
#   - av 包无 Windows 预编译 wheel
# 安装方式：uv sync --extra funasr（仅在 Linux/macOS 上有效）
funasr = [
    "funasr>=1.1.0; sys_platform != 'win32'",
    "modelscope>=1.18.0; sys_platform != 'win32'",
    "torch>=2.0.0; sys_platform != 'win32'",
    "torchaudio>=2.0.0; sys_platform != 'win32'",
]

# 本地 ASR - Whisper（faster-whisper）
# ⚠️ 仅限 Linux/macOS - av 包在 Windows 无预编译 wheel
# 安装方式：uv sync --extra whisper（仅在 Linux/macOS 上有效）
whisper = [
    "faster-whisper>=1.0.0; sys_platform != 'win32'",
    "ctranslate2>=4.0.0; sys_platform != 'win32'",
]

# 本地 ASR - Whisper（原版 OpenAI）- ✅ 全平台支持
# 安装方式：uv sync --extra whisper-openai
whisper-openai = [
    "openai-whisper>=20250625",
    "torch>=2.0.0",
]

# 所有本地推理依赖
# ⚠️ 仅限 Linux/macOS - Windows 请使用 local-windows
# 安装方式：uv sync --extra local
local = [
    # FunASR
    "funasr>=1.1.0; sys_platform != 'win32'",
    "modelscope>=1.18.0; sys_platform != 'win32'",
    "torch>=2.0.0; sys_platform != 'win32'",
    "torchaudio>=2.0.0; sys_platform != 'win32'",
    # faster-whisper
    "faster-whisper>=1.0.0; sys_platform != 'win32'",
    "ctranslate2>=4.0.0; sys_platform != 'win32'",
]

# Windows 本地 ASR（使用 openai-whisper）- ✅ 全平台支持
# 安装方式：uv sync --extra local-windows
local-windows = [
    "openai-whisper>=20250625",
    "torch>=2.0.0",
]

# 开发依赖
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.27.2",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[project.scripts]
voice-server = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# 让 uv 在解析依赖时考虑当前平台
# Windows 上会跳过 FunASR 相关的不兼容依赖
environments = [
    "sys_platform == 'linux'",
    "sys_platform == 'darwin'",
    "sys_platform == 'win32'",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
]
